fix (refs T33506): Rename existing folder "Factory" to prepare move of foundry factories to a folder named "Factory"
